修改 pwm 频率为 uint16_t

+++ user/config/parameters/propeller.h
@@ -9,9 +9,9 @@
 #include "data_define.h"
 
 cJSON* propeller_params_add_to_root(struct propeller_parameters *params);
-void propeller_freq_read_from_root(double *freq, cJSON *node);
+void propeller_freq_read_from_root(uint16_t *freq, cJSON *node);
 void propeller_params_read_from_root(struct propeller_parameters *params, cJSON *node);
-void propeller_params_init_freq(double *params);
+void propeller_params_init_freq(uint16_t *params);
 void propeller_params_init(struct propeller_parameters *params);
 
 void propeller_params_all_init(propeller_t *params);
